requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0

# Statistical analysis
statsmodels>=0.13.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# psutil>=5.8.0  # For memory monitoring
# umap-learn>=0.5.0  # For dimensionality reduction
# hdbscan>=0.8.0  # For alternative clustering
# optuna>=3.0.0  # For hyperparameter optimization

# Development dependencies
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
